☆ star_coords_to_unit_vector and angular_distances functions
- star_coords_to_unit_vector:
The function takes the parameters of the camera (center coordinates, focal length) and an array of star coordinates in the image. I then returns a list of unit vectors for each star

- angular_distances:
The function takes a list of unit vectors and calculates the angular distances between each pair of stars.
Returns an array of all of them

- Also I deleted the two previous digital twins since they were AI and stuff